Finished majority of Create Customer. etc.

+++ MetalLink.Shared/Customers/ProvinceDto.cs
@@ -1,6 +1,6 @@
 public sealed class ProvinceDto
 {
-    public int ProvinceId { get; set; }
+    public long ProvinceId { get; set; }
     public string ProvinceName { get; set; } = string.Empty;
     public string ProvinceCode { get; set; } = string.Empty;
